pyproject.toml
[project]
name = "project-db"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.117.1",
    "langchain>=0.3.27",
    "langchain-community>=0.3.29",
    "langchain-openai>=0.3.33",
    "langgraph-cli[inmem]>=0.4.2",
    "langgraph-sdk>=0.2.6",
    "langgraph[cli]>=0.6.7",
    "ollama>=0.5.4",
    "openai-agents>=0.3.1",
    "pandas>=2.3.2",
    "pip>=25.2",
    "psycopg2-binary>=2.9.10",
    "pydantic>=2.11.9",
    "python-dotenv>=1.1.1",
    "qdrant-client>=1.15.1",
    "tavily-python>=0.7.12",
    "uvicorn[standard]>=0.35.0",
]
